Update MonkRevision.tpa
I thought I posted this earlier, but am not seeing a record of it. If you did receive it and this is a duplicate, I am most sorry and do not intend to inundate you with these. 

The update is to identify if my monk kit is present and update accordingly so it plays nice with your overhaul. Thank you!